Namespace the config utils
